Appease NetBSD (more)

+++ src/hexcom.c
@@ -330,11 +330,11 @@ main (int argc, char **argv)
   /* Flawfinder: ignore */ /* False positive CWE-807/CWE-20 */
   if (getenv ("HEXCOM_NO_PAD") == NULL)
     {
-      unsigned pad_start = first_addr + span;
-      unsigned pad_end = first_addr + (records * RECSZ);
+      size_t pad_start = (size_t)first_addr + span;
+      size_t pad_end = (size_t)first_addr + (size_t)records * RECSZ;
 
       if (pad_end <= ADDRSP && pad_start < pad_end)
-        (void)memset (image + pad_start, 0x1A, (size_t)(pad_end - pad_start));
+        (void)memset (image + pad_start, 0x1A, pad_end - pad_start);
     }
 
   (void)printf ("FIRST ADDRESS %04X\n", first_addr);
